# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateCavelistRequest:
    properties:
      artist_id:
        type: integer
      artist_name:
        type: string
      description:
        type: string
      expiry_promotion:
        description: 'Format: "2006-01-02 15:04:05" atau "2006-01-02T15:04:05Z"'
        type: string
      is_promotion:
        type: boolean
      published_at:
        description: 'Format: "2006-01-02 15:04:05" atau "2006-01-02T15:04:05Z"'
        type: string
      status:
        description: '"draft" atau "publish"'
        type: string
      title:
        type: string
      video_url:
        type: string
    required:
    - artist_id
    - artist_name
    - title
    - video_url
    type: object
...
  handlers.UpdateCavelistRequest:
    properties:
      artist_id:
        type: integer
      artist_name:
        type: string
      description:
        type: string
      expiry_promotion:
        description: 'Format: "2006-01-02 15:04:05" atau "2006-01-02T15:04:05Z"'
        type: string
      is_promotion:
        type: boolean
      published_at:
        description: 'Format: "2006-01-02 15:04:05" atau "2006-01-02T15:04:05Z"'
        type: string
      status:
        description: '"draft" atau "publish"'
        type: string
      title:
        type: string
      video_url:
        type: string
    type: object
...
info:
  contact:
    email: support@soundcave.com
    name: API Support
  description: API documentation for SoundCave Backend - Music streaming platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SoundCave Backend API
  version: "1.0"
paths:
...
  /cavelists:
    get:
      consumes:
      - application/json
      description: Get paginated list of cavelists with filtering and search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by artist_id
        in: query
        name: artist_id
        type: integer
      - description: Filter by status (draft/publish)
        in: query
        name: status
        type: string
      - description: Filter by is_promotion
        in: query
        name: is_promotion
        type: boolean
      - description: Search by title or description
        in: query
        name: search
        type: string
      - default: created_at
        description: Sort field
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all cavelists
      tags:
      - Cavelists
    post:
      consumes:
      - application/json
      description: Create a new cavelist
      parameters:
      - description: Cavelist Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCavelistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create new cavelist
      tags:
      - Cavelists
  /cavelists/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a cavelist by ID
      parameters:
      - description: Cavelist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete cavelist
      tags:
      - Cavelists
    get:
      consumes:
      - application/json
      description: Get a single cavelist by ID
      parameters:
      - description: Cavelist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get cavelist by ID
      tags:
      - Cavelists
    put:
      consumes:
      - application/json
      description: Update cavelist information
      parameters:
      - description: Cavelist ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Cavelist Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCavelistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update cavelist
      tags:
      - Cavelists
  /cavelists/{id}/like:
    post:
      consumes:
      - application/json
      description: Increment likes count for a cavelist
      parameters:
      - description: Cavelist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Increment cavelist likes
      tags:
      - Cavelists
  /cavelists/{id}/share:
    post:
      consumes:
      - application/json
      description: Increment shares count for a cavelist
      parameters:
      - description: Cavelist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Increment cavelist shares
      tags:
      - Cavelists
...
securityDefinitions:
  BearerAuth:
    description: 'To authenticate, provide the API token with the ''Bearer '' prefix.
      Example: ''Bearer your_token_here'''
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"